analytics/requirements.txt
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
databricks-sql-connector==4.1.4
et_xmlfile==2.0.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-pubsub==2.31.1
googleapis-common-protos==1.71.0
grpc-google-iam-v1==0.14.3
grpcio==1.75.1
grpcio-status==1.75.1
idna==3.11
importlib_metadata==8.7.0
lz4==4.4.4
numpy==2.3.4
oauthlib==3.3.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
pandas==2.2.3
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
thrift==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
